+++ src/shared/hacks/rgb-icons/rgb-icons.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "../hacks.hpp"
+
+namespace openhack::hacks {
+
+    /// @brief Kills the player if they die too much in noclip mode
+    class RGBIcons : public EmbeddedHack {
+    public:
+        RGBIcons() : EmbeddedHack("RGB Icons", "rgb_icons") {}
+
+        void onInit() override;
+        void onDraw() override;
+        void update() override;
+        bool isCheating() override { return false; }
+    };
+
+}